+++ index.d.ts
@@ -121,6 +121,8 @@ export interface NavbarProps {
   logo?: ReactElement<any> | string;
   className?: string;
   style?: CSSProperties;
+  onLogoClick?: () => any;
+  logoTitle?: string;
 }
 
 interface NavbarState {
@@ -571,7 +573,6 @@ export type TypographyModifier = 'textPrimary' | 'textDanger' | 'textSuccess' | 
 
 export interface TypographyProps {
   variant: TypographyVariant;
-  description?: boolean;
   children: ReactNode;
   className?: string;
   modifiers?: Array<TypographyModifier>;
@@ -594,6 +595,11 @@ export interface Theme {
       dark?: string;
     };
     text: {
+      main: string;
+      light?: string;
+      disabled?: string;
+    };
+    background: {
       main: string;
       light?: string;
       disabled?: string;
@@ -678,6 +684,7 @@ interface IColors {
   emerald: '#2ecc71';
   peterRiver: '#3498db';
   amethyst: '#9b59b6';
+  asphalt: '#282c35',
   wetAsphalt: '#34495e';
   greenSea: '#16a085';
   nephritis: '#27ae60';
@@ -789,9 +796,6 @@ interface Preview1Props {
   branding?: Branding;
   displayBuyButton?: boolean;
   previewUnavailable?: boolean;
-  minWidth?: string;
-  height?: string;
-  width?: string;
   loginFooterLabel?: string;
   isRestrictedAsset?: boolean;
   handleOpenModal?: (e: any) => any;